+++ Game/TileMap.cpp
@@ -88,10 +88,10 @@ void TileMap::Draw()
 	{
 		for (UINT j = firstcol; j <= lastcol; j++)
 		{
-			float x = tileset_width * (j - firstcol) + Game::GetInstance()->GetCamPosX() - (int)(Game::GetInstance()->GetCamPosX()) % tileset_width;
-			float y = tileset_height * i ;
+			int x = tileset_width * (j - firstcol) + Game::GetInstance()->GetCamPosX() - (int)(Game::GetInstance()->GetCamPosX()) % tileset_width;
+			int y = tileset_height * i ;
 
-			tilemap[i][j]->Draw(1, (int)x, (int)y);
+			tilemap[i][j]->Draw(1, x, y);
 		}
 	}
 }
